refactor(slack-bot): extract neutral slack-blocks module
Move the generic Slack Block Kit primitives (text objects, interactive
elements, layout blocks) out of app-home/slack-types.ts into a neutral
slack-blocks.ts. app-home/slack-types.ts now holds only App-Home view
models and imports the primitives it needs.

Non-App-Home Slack message builders (repo-clarification.ts,
slack-options.ts) no longer reach into app-home/* for these types, which
fixes the dependency direction flagged in the #766 review. Pure type
relocation — no behavior change; typecheck and the 103 slack-bot tests
are unaffected.
